apps/backend/requirements.txt
# Core Dependencies
Flask>=3.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0

# Security & Authentication
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP & Network
requests>=2.31.0
httpx>=0.26.0
websockets>=13.0

# Configuration & Environment
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Type Hints
typing-extensions>=4.9.0

# MQTT & Messaging
paho-mqtt>=1.6.0
gmqtt>=0.7.0

# Graph & Network
networkx>=3.0

# System Monitoring
psutil>=5.9.8

# Vector Database
faiss-cpu>=1.8.0
chromadb>=0.5.0

# Text Processing
sentence-transformers>=3.0.0

# CLI & Display
rich>=13.7.0
click>=8.1.7
tqdm>=4.66.2
pystray>=0.19.0
Pillow>=10.0.0
qrcode>=7.4.0

# Firebase Integration
firebase-admin>=6.5.0

# File Handling
aiofiles>=23.2.0
python-multipart>=0.0.9

# Templates
jinja2>=3.1.0
markupsafe>=2.1.0

# AI Operations System Dependencies
asyncio-mqtt>=0.13.0
prometheus-client>=0.20.0
structlog>=24.1.0
tenacity>=8.2.0

# Web scraping & parsing
beautifulsoup4>=4.12.0
